Trait munge::Destructure

pub unsafe trait Destructure: Sized {
    type Underlying: ?Sized;
    type Destructuring: Destructuring;

    fn underlying(&mut self) -> *mut Self::Underlying;
}

A type that can be destructured into its constituent parts.

Safety

Destructuring must reflect the type of restructuring allowed for the type:

Ref if the type may be restructured by creating disjoint borrows of the fields of Underlying.

Value if the type may be restructured by moving the fields out of the destructured Underlying.

underlying must return a pointer that is non-null, properly aligned, and valid for reads.

Required Associated Types

type Underlying: ?Sized

The underlying type that is destructured.

type Destructuring: Destructuring

The type of destructuring to perform.

Required Methods

fn underlying(&mut self) -> *mut Self::Underlying

Returns a mutable pointer to the underlying type.